Split sync/connection.ts from sync/index.ts
